Fix: 댓글 삭제 확인 alert
